+++ lib/colorus.dart
@@ -4,4 +4,5 @@ export 'src/colorus_grid.dart';
 export 'src/colorus_hue_slider.dart';
 export 'src/colorus_rgb_slider.dart';
 export 'src/colorus_ring.dart';
+export 'src/colorus_slider.dart';
 export 'src/colorus_wheel.dart';