+++ lib/src/colorus_rgb_slider.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-///
-///
+import 'colorus_slider.dart';
+
 class ColorusRGBSlider extends StatelessWidget {
   final Color color;
   final ValueChanged<Color>? onChanged;
@@ -11,60 +11,53 @@ class ColorusRGBSlider extends StatelessWidget {
     super.key,
     required this.color,
     this.onChanged,
-    this.withAlpha = false,
+    this.withAlpha = true,
   });
 
   @override
-  Widget build(BuildContext context) => Column(
-    mainAxisAlignment: MainAxisAlignment.start,
-    mainAxisSize: MainAxisSize.min,
-    children: [
-      if (withAlpha) _buildSlider(true, false, false, false),
-      _buildSlider(false, true, false, false),
-      _buildSlider(false, false, true, false),
-      _buildSlider(false, false, false, true),
-    ],
-  );
+  Widget build(BuildContext context) {
+    return Column(
+      // mainAxisAlignment: MainAxisAlignment.start,
+      mainAxisSize: MainAxisSize.min,
+      crossAxisAlignment: CrossAxisAlignment.stretch,
+      children: [
+        if (withAlpha)
+          _buildSlider(
+            value: color.a,
+            baseColor: color.withValues(alpha: 1.0),
+            // Show color at full opacity in gradient
+            isAlpha: true,
+            label: "A",
+          ),
+        _buildSlider(value: color.r, baseColor: Colors.red, label: "R"),
+        _buildSlider(value: color.g, baseColor: Colors.green, label: "G"),
+        _buildSlider(value: color.b, baseColor: Colors.blue, label: "B"),
+      ],
+    );
+  }
 
-  Widget _buildSlider(bool isA, bool isR, bool isG, bool isB) => Container(
-    height: 30,
-    decoration: BoxDecoration(
-      borderRadius: BorderRadius.circular(15),
-      gradient: LinearGradient(
-        colors: [
-          Colors.white,
-          isA
-              ? color
-              : isR
-              ? Colors.red
-              : isG
-              ? Colors.green
-              : Colors.blue,
-        ],
-      ),
-    ),
-    child: Slider(
-      min: 0,
-      max: 1,
-      onChanged: (onChanged == null)
-          ? null
-          : (v) => onChanged!(
-              Color.from(
-                alpha: isA ? v : color.a,
-                red: isR ? v : color.r,
-                green: isG ? v : color.g,
-                blue: isB ? v : color.b,
-              ),
-            ),
-      overlayColor: WidgetStatePropertyAll(Colors.black),
-      thumbColor: Colors.white,
-      value: isA
-          ? color.a
-          : isR
-          ? color.r
-          : isG
-          ? color.g
-          : color.b,
-    ),
-  );
+  Widget _buildSlider({
+    required double value,
+    required Color baseColor,
+    required String label,
+    bool isAlpha = false,
+  }) {
+    return ColorusSlider(
+      value: value,
+      baseColor: baseColor,
+      withCheckerBoard: isAlpha,
+      orientation: Orientation.landscape,
+      onChanged: (v) {
+        if (onChanged == null) return;
+        onChanged!(
+          Color.from(
+            alpha: isAlpha ? v : color.a,
+            red: label == "R" ? v : color.r,
+            green: label == "G" ? v : color.g,
+            blue: label == "B" ? v : color.b,
+          ),
+        );
+      },
+    );
+  }
 }